Decouple player view calculation from Server's Master
